+++ src/isomedia/isom_read.c
@@ -4878,7 +4878,7 @@ GF_Err gf_isom_get_file_offset_for_time(GF_ISOFile *movie, Double start_time, u6
 		return GF_BAD_PARAM;
 
 	if (!movie->main_sidx) return GF_NOT_SUPPORTED;
-	start_ts = start_time * movie->main_sidx->timescale;
+	start_ts = (u64) (start_time * movie->main_sidx->timescale);
 	cur_start_time = 0;
 	offset = movie->main_sidx->first_offset + movie->main_sidx_end_pos;
